+++ sdk/python/pulumi_azure/datafactory/__init__.py
@@ -7,6 +7,7 @@ from .factory import *
 from .dataset_mysql import *
 from .dataset_postgresql import *
 from .dataset_sql_server_table import *
+from .linked_service_data_lake_storage_gen2 import *
 from .linked_service_mysql import *
 from .linked_service_postgresql import *
 from .linked_service_sql_server import *